Safari-609.1.11.1/LayoutTests/svg/dom/SVGTransformList-expected.txt

This test checks the SVGTransformList API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check passing invalid arguments to 'createSVGTransformFromMatrix'
PASS transform.createSVGTransformFromMatrix('aString') threw exception TypeError: Type error.
PASS transform.createSVGTransformFromMatrix(1) threw exception TypeError: Type error.
PASS transform.createSVGTransformFromMatrix(true) threw exception TypeError: Type error.
Check passing valid arguments to 'createSVGTransformFromMatrix'
PASS transform.createSVGTransformFromMatrix(null) did not throw exception.
PASS transform.createSVGTransformFromMatrix(svgElement.createSVGTransform()) did not throw exception.
PASS transform.createSVGTransformFromMatrix(svgElement) did not throw exception.
PASS transform.createSVGTransformFromMatrix(undefined) did not throw exception.
PASS transform.createSVGTransformFromMatrix() did not throw exception.
PASS t.matrix.a is 2
PASS t.matrix.b is 0
PASS t.matrix.c is 0
PASS t.matrix.d is 3
PASS t.matrix.e is 10
PASS t.matrix.f is 20
PASS t.matrix.a is 3
PASS t.matrix.b is 0
PASS t.matrix.c is 0
PASS t.matrix.d is 4
PASS t.matrix.e is 0
PASS t.matrix.f is 0
PASS t.matrix.a is 1
PASS t.matrix.b is 0
PASS t.matrix.c is 0
PASS t.matrix.d is 1
PASS t.matrix.e is 0
PASS t.matrix.f is 0
PASS successfullyParsed is true

TEST COMPLETE
